Move company detail page read from Supabase to Typesense
The company detail page issued a SQL query (LATERAL join across
company / industry / company_description / industry_name) on every cache
miss — ~10% of Supabase compute over 41.8k calls. The 600s Redis cache
in front of getCompanyBySlug was already there; the long-tail cold-key
distribution kept Supabase load high.

This PR moves the read to Typesense, the same in-memory layer that
already served the page's active_posting_count call, and treats Postgres
as fallback. The two prior round-trips (SQL + Typesense) collapse to
one (Typesense only).

Schema changes (in `apps/crawler/src/typesense_schema.py`, new module
extracted from `scripts/typesense-setup.py`):
- `company` collection gains `logo`, `website`, `employee_count_range`,
  `founded_year`, plus per-locale `description_{de,fr,it}` and
  `industry_name_{de,fr,it}` (English stays in `description` /
  `industry_name`; reader falls back to English when a locale variant
  is missing).
- All new fields are `optional: true` and `index: false` so they fit
  Typesense's in-place PATCH semantics on alter.
- `setup-typesense` is now a `crawler` CLI subcommand. Setup is
  idempotent on every run (creates missing + PATCHes existing schemas
  with new fields). Wired into `deploy.sh` between Alembic and `crawler
  sync` so a deploy that adds fields ships safely.

Sync (`sync_companies_typesense`) now fetches all locales of
`company_description` and `industry_name`, plus the new scalar fields
from the `company` table, and writes the locale-suffixed fields. Also
fixes a pre-existing gap: the English description now falls back to
`c.description` when no `company_description['en']` row exists,
matching the Postgres path semantics.

Web (`getCompanyBySlug`):
- Primary: Typesense filter-by-slug, locale-aware field picker.
- Fallback (Typesense exception OR 0 hits): existing Postgres SQL.
  Brand-new companies whose Typesense upsert lagged still render.
  Bot 404s pay a cheap PK lookup; cache layer skips storing nulls so
  nonexistent slugs aren't poisoned.
- `activeJobCount` on the Postgres fallback returns 0 — only the header
  count is affected; the postings list itself is a separate Typesense
  call (getCompanyPostings).

Docs updated: `docs/11-typesense.md` gains a Read paths summary;
`docs/07-system-design.md` corrects the sync-targets count (3 → 4);
`docs/13-seo-and-indexnow.md` and root `AGENTS.md` reflect the new read
path.
