+++ AGENTS.md
@@ -88,9 +88,9 @@ Developer guidance for agent reasoning style lives in [docs/agents.md](docs/agen
 
 ## Typesense (Search Engine)
 
-All search, typeahead, browse-all modals, and watchlist search are served by Typesense. Supabase Postgres still handles non-search reads (posting detail, user data).
+All search, typeahead, browse-all modals, watchlist search, and the **company detail page** are served by Typesense. Supabase Postgres still handles posting detail (full description blob), user/auth data, watchlist mutations, and acts as a graceful fallback when Typesense is unreachable.
 
-See [docs/11-typesense.md](docs/11-typesense.md) for full deployment details.
+See [docs/11-typesense.md](docs/11-typesense.md) for full deployment details, including the read-paths summary.
 
 ### Infrastructure
 
@@ -122,8 +122,13 @@ Key design choices:
 
 ### Collection Management
 
+Schema source of truth: `apps/crawler/src/typesense_schema.py`. Setup is idempotent — it creates missing collections + aliases AND patches existing collections in-place to add any new fields (no rebuild required). Runs automatically on every crawler deploy via `deploy.sh` before `crawler sync`.
+
 ```bash
-# Create or recreate collections (from apps/crawler/)
+# Idempotent create + patch (from inside the crawler image)
+uv run crawler setup-typesense [--force]
+
+# Operator-facing wrapper (dev workflows)
 cd apps/crawler && uv run python ../../scripts/typesense-setup.py [--force]
 
 # Full re-index from Postgres
@@ -136,13 +141,13 @@ cd apps/crawler && uv run python ../../scripts/typesense-backfill-local.py [--li
 ### Indexing Pipeline
 
 - **Exporter** (CDC): two-cursor design — Supabase and Typesense cursors advance independently. Concurrent upserts via `asyncio.gather`
-- **Sync**: taxonomy collections (location, occupation, seniority, technology, company) populated after CSV sync. Handles taxonomy rename detection
+- **Sync**: taxonomy collections (location, occupation, seniority, technology) and the `company` collection populated after CSV sync. Company docs include extended fields (logo, website, employee_count_range, founded_year) and per-locale variants (`description_{de,fr,it}`, `industry_name_{de,fr,it}`) for the company detail page reader. Handles taxonomy rename detection
 - **Reconciliation**: daily count check + sample comparison
 - **refresh-typesense**: periodic count refresh for taxonomy/company collections + watchlist reconciliation
 
 ### Web App Integration
 
-`TypesenseSearchProvider` replaces `PostgresSearchProvider` (one-shot cutover). Graceful degradation: all errors return empty results, Postgres fallback for watchlist write functions. No Redis cache on main search (Typesense is fast enough); cached for unfiltered homepage (60s) and popular watchlists (120s).
+`TypesenseSearchProvider` replaces `PostgresSearchProvider` (one-shot cutover). The company detail page (`getCompanyBySlug`) reads from the `company` collection, falling back to Supabase on Typesense error or 0 hits. Graceful degradation: all errors return empty results, Postgres fallback for watchlist write functions. No Redis cache on main search (Typesense is fast enough); cached for unfiltered homepage (60s), popular watchlists (120s), and company detail (`ttl: 600`, skip-null to avoid poisoning brand-new slugs).
 
 ## SEO and IndexNow
